Separate RPC wallet initilization and connection
Add RPC handlers for explicitly creating a wallet that returns an error if the
wallet already exists, and for connecting to an existing wallet that errors if
the wallet does not already exist. This allows us to build interfaces that rely
on the RPC service with a more predictable user experience.