Remove protoactor-go version constraint
We believe the behavior we saw that motivated adding this in the first
place was actually a red herring; an intermittent bug that happened to
occur before this change was made and then did not re-occr after it was
made for two of us.

It should (hopefully) just use the version that the client lib brings
in.