Revert serialized signatures signers to bool array
This is a better language-agnostic encoding of this data structure. I think the arrays might end up being longer than the number of signers due to the way the bitarray lib works, but the extras will all be "false" so maybe that's OK as long as we don't rely on its cardinality to tell us anything about the size of the notary group.